+++ functions/ft_calloc.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   ft_calloc.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: josmarqu <josmarqu@student.42madrid.com    +#+  +:+       +#+        */
+/*   By: josmarqu <josmarqu@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/01 09:20:35 by josmarqu          #+#    #+#             */
-/*   Updated: 2023/02/01 09:20:44 by josmarqu         ###   ########.fr       */
+/*   Updated: 2023/02/05 18:41:50 by josmarqu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stdlib.h>
 
-void	*ft_calloc(int nmemb, int size)
+void	*ft_calloc(int nmemb, size_t size)
 {
 	return (malloc(nmemb * size));
 }
