+++ functions/ft_strlcat.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strlcat.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: josmarqu <josmarqu@student.42madrid.com    +#+  +:+       +#+        */
+/*   By: josmarqu <josmarqu@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/01 09:25:19 by josmarqu          #+#    #+#             */
-/*   Updated: 2023/02/01 09:25:22 by josmarqu         ###   ########.fr       */
+/*   Updated: 2023/02/05 18:43:40 by josmarqu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 int	ft_strlen(const char *s);
 
-int	ft_strlcat(char *dest, const char *src, int destsize)
+int	ft_strlcat(char *dest, const char *src, size_t destsize)
 {
 	int				dest_len;
 	unsigned int	len;
